modified the way names are encoded
The blank space at the end of each name is now an extra dimension in the one-hot vectors (as opposed to being represented by empty vectors). In addition, the raw names are capitalized, which I had mistakenly not accounted for, causing the first letter of every name to be percieved as a space.